Catch error connecting to AdobeConnect
When the calendar_event api converts an event with a webconference
to json, it tries to retrieve recordings. In the case of AdobeConnect
it has to connect to the service. If that fails (like it did for the
referenced support week), it kills the entire api response. This is a
little change to catch the error where it happens so the api can return
everything but the recordings it didn't get.

closes LS-3021

test plan:
  - from /plugins/adobe_connect
    - enable the plugin
    - set the domain to anything
    - set the meeting folder URL to anything
    - set everything else to anything
    - Apply
  - in the calendar, create an event in a course
  - click the Add Conferencing button and add Adobe Connect
  - refresh the calendar page
  > expect the event to be in the api response and in the calendar